requirements.txt
notebook

# Основные зависимости
langchain
langchain-gigachat
langchain-community

# Обработка документов
unstructured[pdf,docx]  # Главная библиотека для парсинга
pymupdf               # Альтернатива для PDF
pdf2image             # Для конвертации PDF в изображения
pillow                 # Работа с изображениями
easyocr                  # OCR для сканов
python-magic           # Определение типа файла
python-magic-bin
langdetect

# Работа с данными
pydantic
pandas
numpy==1.24.3

# Векторная БД
weaviate-client

# Деплой и API
fastapi
uvicorn[standard]
streamlit
python-multipart

# Утилиты
python-dotenv
tqdm
loguru